+++ packages/common/src/utils/filters.ts
@@ -57,6 +57,19 @@ export const getTotalFilterRules = (filters: Filters): FilterRule[] => [
 export const countTotalFilterRules = (filters: Filters): number =>
     getTotalFilterRules(filters).length;
 
+export const hasNestedGroups = (filters: Filters): boolean => {
+    const hasGroups = (filterGroup: FilterGroup): boolean => {
+        const items = isAndFilterGroup(filterGroup)
+            ? filterGroup.and
+            : filterGroup.or;
+        return items.some(isFilterGroup);
+    };
+    return (
+        (!!filters.dimensions && hasGroups(filters.dimensions)) ||
+        (!!filters.metrics && hasGroups(filters.metrics))
+    );
+};
+
 export const getItemsFromFilterGroup = (
     filterGroup: FilterGroup | undefined,
 ): FilterGroupItem[] => {